pyproject.toml
[tool.poetry]
name = "JWT-Auth"
version = "2.0.0"
description = "Example of an application with JWT authorization"
authors = ["befunny <befunny@doubletapp.ai>"]

[tool.poetry.dependencies]
python = "==3.10.7"
fastapi = "==0.82.0"
uvicorn = "==0.18.3"
asgi-lifespan = "==1.0.1"
httpx = "==0.23.0"
environs = "==9.5.0"
pyjwt = "==2.4.0"
asyncpg = "==0.26.0"
tortoise-orm = "==0.19.2"
pytest = "==7.1.3"
Faker = "==15.1.0"

[build-system]
requires = ["poetry-core>=1.1.13"]
build-backend = "poetry.core.masonry.api"
